+++ packages/core/styles/src/components/ThemeProvider/contexts/ThemeProviderContext/ThemeProviderContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { ThemeProviderContextProps } from "./themeProviderContext.types";
+
+export const ThemeProviderContext = createContext<ThemeProviderContextProps>(
+  {} as ThemeProviderContextProps
+);