+++ packages/core/styles/src/components/ThemeProvider/contexts/ThemeProviderContext/themeProviderContext.types.ts
@@ -1,0 +1,7 @@
+import { MutableRefObject } from "react";
+import { Theme } from "../../themeProvider.types";
+
+export interface ThemeProviderContextProps {
+  refThemeProvider: MutableRefObject<null | HTMLDivElement>;
+  currentTheme: Theme;
+}