+++ packages/core/styles/src/components/ThemeProvider/index.ts
@@ -1,5 +1,6 @@
 import { ThemeProvider } from "./ThemeProvider";
 
 export { ThemeProvider } from "./ThemeProvider";
+export { useTheme } from "./hooks";
 export type { ThemeProviderProps } from "./themeProvider.types";
 export default ThemeProvider;