+++ packages/react/src/atomic/MultiSelect/src/components/MultiSelectSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { MultiSelectSkeleton } from "./MultiSelectSkeleton";
+
+export { MultiSelectSkeleton } from "./MultiSelectSkeleton";
+export default MultiSelectSkeleton;