+++ packages/react/src/atomic/MultiSelect/src/components/MultiSelectSkeleton/multiSelectSkeleton.stories.tsx
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { MultiSelect } from "../../MultiSelect";
+
+const meta: Meta<typeof MultiSelect.Skeleton> = {
+  title: "atomic/MultiSelect/MultiSelect.Skeleton",
+  component: MultiSelect.Skeleton,
+  tags: ["autodocs"],
+};
+
+export default meta;
+type Story = StoryObj<typeof MultiSelect.Skeleton>;
+
+export const basic: Story = { args: {} };