+++ packages/react/src/atomic/MultiSelect/src/components/MultiSelectSkeleton/multiSelectSkeleton.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import { SkeletonProps } from "@nimbus-ds/skeleton";
+
+export type MultiSelectSkeletonSkeletonProperties = Partial<
+  Pick<SkeletonProps, "width" | "data-testid">
+>;
+
+export type MultiSelectSkeletonSkeletonProps =
+  MultiSelectSkeletonSkeletonProperties & HTMLAttributes<HTMLDivElement>;