+++ packages/react/src/atomic/Popover/src/Popover.tsx
@@ -15,7 +15,7 @@ import {
   arrow as arrowUI,
   offset as offsetUI,
 } from "@floating-ui/react";
-import { popover } from "@nimbus-ds/styles";
+import { popover, useTheme } from "@nimbus-ds/styles";
 
 import { PopoverProps } from "./popover.types";
 
@@ -49,7 +49,7 @@ const Popover: React.FC<PopoverProps> = ({
   const arrowRef = useRef(null);
   const [isVisible, setVisibility] = useState(false);
 
-  const open = useMemo(
+  const open: boolean = useMemo(
     () => (visible === undefined ? isVisible : visible),
     [visible, isVisible]
   );
@@ -74,6 +74,8 @@ const Popover: React.FC<PopoverProps> = ({
         },
       }),
   ].filter((middleware) => middleware !== false);
+
+  const { refThemeProvider } = useTheme();
 
   const { context, floatingStyles } = useFloating({
     open,
@@ -117,7 +119,10 @@ const Popover: React.FC<PopoverProps> = ({
             })
           : children}
       </div>
-      <FloatingPortal id="nimbus-popover-floating">
+      <FloatingPortal
+        id="nimbus-popover-floating"
+        root={refThemeProvider?.current}
+      >
         {open && (
           <div
             {...otherProps}